fixed predict api
